requirements.txt
astral==3.2
certifi==2026.1.4
charset-normalizer==3.4.4
customtkinter==5.2.2
darkdetect==0.8.0
geographiclib==2.1
geopy==2.4.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.2
packaging==26.0
pandas==3.0.1
prettytable==3.17.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rich==14.3.3
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
